refactor(no_intersection): delete the adaptive worms, re-parent FreeTargetWorm
AdaptiveIntersectionWorm and TwoLinkAdaptiveWorm both jam: at kappa = 0.01
they have no legal move in 97.7% and 94.8% of proposals respectively, with a
median clean set of zero.  FreeTargetWorm absorbs the four helpers it actually
used and now subclasses IntersectionWorm directly.
